fix/removendo-restantes

+++ src/main/java/com/projeto/sistema_epi/repository/EmprestimoRepository.java
@@ -3,5 +3,5 @@ package com.projeto.sistema_epi.repository;
 import com.projeto.sistema_epi.entity.EmprestimoEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmprestimoRepository extends JpaRepository<EmprestimoEntity,Long> {
+public interface EmprestimoRepository extends JpaRepository<EmprestimoEntity,Integer> {
 }
